+++ CHANGELOG.md
@@ -1,12 +1,15 @@
 ## [Unreleased]
 
 ### Added
-
 - google resources: `google_compute_backend_bucket`, `google_project_iam_custom_role`, `google_storage_bucket_iam_policy`, `google_compute_instance_iam_policy`
   ([PR #97](https://github.com/cycloidio/terracognita/pull/97))
+- aws: `aws_lb`, `aws_lb_listener`, `aws_lb_listener_rule`, `aws_lb_target_group`
+  ([PR #96](https://github.com/cycloidio/terracognita/pull/96))
 
 ### Changed
 
+- filter: update IsExcluded and add IsIncluded to verify multiple resources.
+  ([PR #96](https://github.com/cycloidio/terracognita/pull/96))
 - Provide filters to resource functions instead of tags only
   ([PR #92](https://github.com/cycloidio/terracognita/pull/92))